Testing failures for C172

+++ Modules/FailuresModule/Model/App/Percentage.cs
@@ -27,7 +27,7 @@ namespace FailuresModule.Model.App
 
         public override string ToString()
         {
-            return (int)(value * 100) + "%";
+            return (value * 100) + "%";
         }
     }
 }
